refactor(tui): extract prompt config, keybindings, and completion menus
