+++ src/pythinker_code/ui/shell/prompting/__init__.py
@@ -1,6 +1,11 @@
 """Deep prompt rendering and session-resource modules."""
 
 from pythinker_code.ui.shell.prompting.clipboard import ClipboardAdapter
+from pythinker_code.ui.shell.prompting.config import (
+    BgTaskCounts,
+    PromptConfig,
+    PromptProviders,
+)
 from pythinker_code.ui.shell.prompting.footer import (
     FooterContent,
     FooterViewModel,
@@ -21,6 +26,10 @@ from pythinker_code.ui.shell.prompting.history import (
     PromptHistoryStatus,
     PromptHistoryStore,
 )
+from pythinker_code.ui.shell.prompting.keybindings import (
+    PromptController,
+    build_prompt_key_bindings,
+)
 from pythinker_code.ui.shell.prompting.renderer import (
     PromptSceneAllocation,
     PromptSceneBudget,
@@ -29,12 +38,16 @@ from pythinker_code.ui.shell.prompting.renderer import (
 from pythinker_code.ui.shell.prompting.toasts import ToastManager, ToastSnapshot
 
 __all__ = (
+    "BgTaskCounts",
     "ClipboardAdapter",
     "FrozenFragments",
     "FooterContent",
     "FooterViewModel",
     "GitSnapshot",
     "GitStatusIndex",
+    "PromptConfig",
+    "PromptController",
+    "PromptProviders",
     "PromptFrame",
     "PromptFrameCollector",
     "PromptHistoryError",
@@ -46,6 +59,7 @@ __all__ = (
     "ToastSnapshot",
     "allocate_prompt_scene_rows",
     "background_task_summary",
+    "build_prompt_key_bindings",
     "freeze_fragments",
     "select_footer_content",
     "truncate_footer_left",
